add: use tanstack query in vocabulary resource
Rewriting the logic and tests wasn't too hard apart from one blunder on
my side.

So far it's more code but it already helped me solve couple of issues.
The first was that dashboard recent vocabularies were always refetched
on each visit. My plan was to introduce another resource but this
resource would have to be centrally managed if I didn't want to refetch
on each visit again. And then I'd have to keep it in sync with the other
resources of course. Instead of this toil I just declare the recent
vocabularies query as invalidated when a vocabulary gets updated and
that's all! Thanks to the query cache, the refetch is done only on
demand. I'd say this is worth it. Of course if I wanted I could decouple
this architecture even a more by registering query keys which need to be
updated on changes but I'll skip that part for now.

I also got aware of Suspense so I immediately started using it.

I got rid of the query.ts file since it was kinda hidden dependency that
would cause inconsistencies in tests due to 2 query clients being created.

Can't wait to convert the vocabularies resource!