add: use tanstack for vocabularies instead of resource
I finally figured out how it works with the query keys. Let's take
recent and all vocabularies queries. I though each needs a different
query key, but it turns out that using the same query one is much
simpler. Not only can I make both refetch or invalidate in one go, which
is actually mostly what I want, but it also foregoes multitude of
different query keys. The only thing I do to differentiate is specifying
a different type for the recent query.

I learned in the meantime, that it's confusing that the `queryKey` is
an array because I thought I can pass multiple keys in this array to eg.
invalidate multiple caches, but that's not how it works, and the array
only ever specifies one key, kinda like with angular route paths.

In this commit I realized how many times I forgot to sync recent and all
vocabularies with vocabulary resource changes, or just the former two
between each other. In order to illustrate how important this is, imagine
this case: If I used tanstack query before, had the invalidation
logic in place for vocabularies and decided to add recent vocabularies
to the app, I just would have to use `createQuery` with the same key and
that would be it! When 'vocabularies' would invalidate, the recent ones
would invalidate too unless I specifically excluded them. And it makes
sense, it's the same resource just with a different query. Now without
tanstack, I'd have to:
- create a new resource for recent vocabularies
- make that resource independent of a component in order to cache
- create a small API to keep it in sync with other resources, eg. when
I'd update a vocabulary so that it ends up first in the recent
vocabularies
- create some kind of a signal based logic that would invalidate the
resource.
- add it into the app init and test app init logic

That's a lot of stuff for one simple resource! So while tanstack means
nominally more code, it seems to actually save code in the caching logic
a lot and more importantly to forego problems and lot of work.