fix: do not update vocabulary after every word tested
This was caused by updating the savedProgress on the vocabulary after
each word guess. That's fine, but it lead to reactive triggering of the
effect that was updating the vocabulary's updated_at prop. This is not
desired and so I switched from pure reactivness to checking whether the
vocabulary id changed.

This bug wasn't caused by the switch to tanstack, but rather by updating
the savedProgress directly on the vocabulary.