fix: confirm dialog duplicate button
The thing is that DialogTrigger is itself a button so when I passed
another button in there, I ended up with nested buttons. The problem
with that was also that the styles applied to the inner weren't applied
to the outer, so it eg. didn't get fully hidden and users could still
click it by accident.

So I made a span out of the outer and also make it possible to pass a
class into it. Span wrapping a button is valid HTML and this solution
prevents an akward passing of props onto the passed-in button, which
proved inpracticat or in certain cases impossible.