add: basic word guessing breakdown
After considering a horizontal bar chart, I settled for a clean looking
table. The reason is that the number of invalid attempts and the fact
that user finished the word don't really belong on the same axis.

I think down the line it will make sense to connect test results with
words because for the results page, neither makes sense alone. Whether I
do this in some resource or in the top component is the question.