fix: prevent tester from needlessly re-rendering
This was a weird one. Let me start with a description of the problem.
When I navigated from some page to vocabulary tester page, wrote the
first guess and went to the next word, the input regained focus for a
moment but then lost it immediately. If I were to reload the tester page
and try the same thing, it would work as expected. In any case,
subsequent guesses would not have this problem. This is already weird
enough.

Investigating the problem was difficult since to find the culprit of
unexpected blur is never straight forward, and indeed there was no
other focus triggered or anything that obviously cause it. Then I noticed
in the devtools that when I went to the next word for the 1st time after
the navigation, the whole page component was briefly removed from DOM and
then immediately re-inserted. This was the cause for the lost focus and
it became evident very soon that `<Suspense>` was to blame.

But was it? First of all Suspense only reacts to the resource being in
the loading state. So I had a look what's causing the loading and found
that it's the saved progress being updated and propagated back to the
tanstack query which I use on the test page. Indeed, when I removed this
query update - `setQueryData` - the problem was gone. That's nice, but
this update doesn't cause a refetch and is synchronous. Still, I thought
maybe the query object's state changes and that might trigger the
Suspense fallback. But when I logged the query state, it was always
'success' and never 'loading'.

At this point it became clear it's a quirk somewhere between Tanstack
and Solid. Being the detective I am, I started debugging the issue but
even after an hour didn't find anything that'd give me any guidance. So
in the end I decided to just use `<Show>` instead of `<Suspense>` and
voila, the problem went away. It's a nice workaround but I have to stay
vigilant for the other pages and their Suspenses as this feels like it
could happen elsewhere too.