change: the result system
So previously the result could only have been a de-facto boolean, which
doesn't suite the guess mode aka Anki mode well since the outcome can
have multiple degrees of correctness - I so far went for 4.

The other problem was that the information about the status and the end
result was all over the place, so I put all this info into 2 properties
- status and result. It's a little wordier but imo much cleaner.

This commit only provides backwards compatibility for the write mode
results but doesn't offer any display nor storage for the guess mode
results. This will be done in the next commit.