add: guess result recording
It gets saved into the progress and test result, however the thing that
doesn't get saved yet is the attempts history. As it turns out, storing
just the number of invalid attempts is not enough, and I should store
the full attempt results history. Not sure how I will do it yet but
I'll probably keep it simple and just store it as JSON in the DB.