fix(perf): use dict to optimize performance of saved progress
Previously, when I wanted to display selected words on the test page, I
had to iterate through them and check for each if it's present in the
saved progress words. This was very slow when there are many words.

Now, I'm using a dictionary to use hash table access which is O(1).

The only problem is that the VocabularyTester is still iterating the old
way but I won't bother with it just yet as the performance is good enough.