change: consider Ok guesses as invalid
I started bumping into the situation where I almost guessed the word but
not completely, yet I wanted to repeat it to get it right. However,
pressing Ok would remove it from the queue and I wouldn't be able to
repeat it. This change removes only the completely correct guesses.