fix: test progress with deleted words
Deleting a word would previously not update the test progress correctly,
which resulted in deleted word still being referenced in the test
progress, which broke the test. So I have to check for this specifically
since I already have some saved progress with this invalid state and I
prevent this from happening by immediately updating the test results.