add: make results selectable
For that purpose I extracted the selection logic into a separate util,
which is akin to a React hook.

The thing pissing me off is that I have to use find() to find out if a
checkbox is checked. This should be easier and maybe I could open a PR
for this.