change: use done instead of status
Status was capturing too much info and had too much variation. This made
it complicated when retrieving eg. color when combined with result. So
I now capture a little more in the result and less in the `done`, which
replaced status.

I also made the vocab test progress/result a little prettier.