refactor: change navigation to test
Previously, the navigation to test wasn't idempotent. Meaning the words
ids in the URL would on each refresh create a new test. This lead to an
awkward bug where a previous test would be lost if I did a refresh.

So I changed it to get rid of the words ids when the test initializes.
When it does, and the test doesn't exist, it creates a new one. I also
got rid of the "useSavedProgress" since it's not needed with test IDs
in the URL - the test itself knows if it's in progress or not.

I also did a nice refactoring where I use constants for route paths. This
proved to be useful since changing urls in the whole codebase was a pain
with hardcoded strings.

Last but not least, I simplified the tester logic. I don't need the weird
non-state "isvalid" variable and I don't split the validation logic into
multiple functions anymore.