fix: things around last test progress
Firstly I was using a wrong tanstack query key for it. Secondly I fixed the annoying "Invalid date" display.

Thirdly, I update the query from a component isntead from a resource. I find that some queries are so specific that it would be weird to update them from a generic resource. To make this more encapsulated, I started using mutations.