fix: conjugations view
Here's the deal. I've been using mlconjug3 for a while but it was never
perfect. It was super slow on cold start and it wasn't very faster after
it either. Also, it was hosted on pythonanywhere which was horrible to
maintain on the free plan.

The other argument is Why use ML for conjugations? mlconjug3 was making
mistakes, didn't have such a large database and consumed too many
resources, to train at least. And it also supported only Romance
languages. All that in a world where we have large, up-to-date word DBs.

So I began thinking of using Wiktionary, which even has REST API for this,
but it's hard to parse conjugations from its HTML. Each language has a
different table structure.

So I resorted to using Verbix, since I discovered that:
- They made the needed API key accessible in their JS file through their
docs.
- They have a REST API that returns conjugations in JSON format!
- Their license requires only attribution and non-commercial use, which
is fine for this project.
So easy!

Then I only needed to adapt the Conjugations view to this new data source.