add: input reset button to country select
The icon sat misaligned in the combobox until I found that it's due to
button styles which were lifting the icons up. This was ridiculous since
the icons are always rectangles and sure enough, I found the button was
using `items-end` instead of centering.

This allowed me to remove the button.module.css file, yay!