fix: prevent archived words from returning in tests
They were returning because I wasn't filtering them when the test loaded
from a saved test result. However, I don't think additional filter   is
the solution, because it might happen I will want to test archived words.

That's why I decided to give an automatic "passed" note when word is
archived. This is not only logical, because the user is most probably
already familiar with it, but also practical since it prevents the word
from reappaering in the test, yay!