build: simplify service worker build setup
First of all, removes the need to add inline lib references by specifying
the libs in tsconfig.

Secondly, I moved the tsconfig together with the SW
file into a separate folder in order to separate it from the rest of the
project. This also lead to a more straightforward setup in the sub
tsconfig.